Finalize login
