+++ package.json
@@ -4,6 +4,7 @@
   "version": "0.1.1",
   "private": true,
   "dependencies": {
+    "@hookform/resolvers": "^1.1.1",
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.56",
@@ -22,6 +23,7 @@
     "react": "^16.14.0",
     "react-axios": "^2.0.3",
     "react-dom": "^16.14.0",
+    "react-hook-form": "^6.12.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^4.0.0-next.98",
     "typescript": "~3.7.2"
